Calculate CV of inter-cell type difference

This function calculates the coefficient of variation of pairwise Jensen-Shannon distances between cell types. High values indicate that certain cell type pairs exhibit much more dramatic differences in isoform usage than others, suggesting specialised functional distinctions between specific cell populations.

Usage

.calculate_inter_cell_type_difference_variability(specificity_result)

Arguments

specificity_result

Result from .calculate_inter_cell_type_specificity()

Value

Coefficient of variation of cell type difference distances