Blood dataset gene expression counts

Gene-level expression count matrix from blood single-cell RNA-seq data. This data is derived from Wang et al. (2022) study on alternative splicing during hematopoietic stem cell formation.

Format

A data frame

Source

Wang et al. (2022). Science Advances 8(1):eabg5369. Data from https://zenodo.org/records/5706781. Licensed under CC BY 4.0.

References

Wang F, et al. (2022). Single-cell architecture and functional requirement of alternative splicing during hematopoietic stem cell formation. Science Advances 8(1):eabg5369.